+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+* Hide the system bars in the full screen viewer
+
 ## [v1.37.0] - 2026-01-21
 [v1.37.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.37.0